% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.duration.R
\docType{methods}
\name{as.duration,character-method}
\alias{as.duration,character-method}
\title{as.duration.character}
\usage{
\S4method{as.duration}{character}(x)
}
\arguments{
\item{x}{character to convert}
}
\description{
coerce object to a \code{lubridate} duration object
}
\examples{
  as.duration( "7.25 days" )
  as.duration( "3600 seconds" )
  as.duration( "1 week" )

}
