% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-differences.R
\name{diffdays}
\alias{diffdays}
\alias{diffweeks}
\title{Difference between dates expressed in days or weeks}
\usage{
diffdays(x, y)

diffweeks(x, y)
}
\arguments{
\item{x}{the first date/time object}

\item{y}{the second date/time object}
}
\description{
Difference between dates expressed in days or weeks
}

