% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed-months.R
\name{elapsed_months}
\alias{elapsed_months}
\title{Calculate elapsed months between days}
\usage{
elapsed_months(end_date, start_date)
}
\arguments{
\item{end_date}{object coercible to POSIXlt}

\item{start_date}{object coercible to POSIXlt}
}
\value{
integer
}
\description{
Calculate elapsed months between days
}
\examples{

 elapsed_months("2016-03-10", "2016-01-01")



}
\references{
http://stackoverflow.com/a/26640698
}

