% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_yyyy.R
\name{mmm_yyyy}
\alias{mmm_yyyy}
\title{Date Parsing}
\usage{
mmm_yyyy(x, ...)
}
\arguments{
\item{x}{character}
}
\description{
Parse Mon-Year (i.e. "Jan 1970") format
}
\details{
Parses MMM YYYY format; the parsed date assumes the first moment of the
period.

An abbreviated name for this function cannot be used since it would be "by"
which is a common reserved word
}
