% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin.R
\name{origin}
\alias{origin}
\title{Get origin for a time series class}
\usage{
origin(x, ...)
}
\arguments{
\item{x}{time series object}

\item{...}{addition arguments}

\item{tzone}{string; time zone abbreviation}
}
\value{
A one element vector with the class of \code{x}, denoting the
}
\description{
Return the origin for a time-series class
}
\seealso{
\code{\link{origins}}
}
