% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origins.R
\docType{data}
\name{epoch_origin}
\alias{epoch_origin}
\alias{origin_Date}
\alias{origin_POSIXct}
\alias{origin_POSIXlt}
\alias{origin_chron}
\alias{origin_epoch}
\title{1970-01-01 UTC}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.}
\usage{
epoch_origin

origin_epoch

origin_POSIXct

origin_POSIXlt

origin_chron

origin_Date
}
\description{
Origin is the date-time for 1970-01-01 UTC in POSIXct format. This date-time
is the origin for the numbering system used by POSIXct, POSIXlt, chron, and
Date classes.
}
\details{
This is a copy of \code{\link[lubridate]{origin}}.
}
\examples{
  epoch_origin

  origin_epoch
  origin_POSIXct

}
\seealso{
lubridate::origin
}
\keyword{chron}
\keyword{data}
\keyword{datasets}

