% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week52.R
\name{week52}
\alias{week52}
\title{week52}
\usage{
week52(x)
}
\arguments{
\item{x}{a data-time object; Must be a POSIXct, POSIXlt, Date, chron,
yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or fts object.}
}
\value{
integer
}
\description{
Get/Set week of a date-time with consistent 52 week years
}
\details{
Same as week, but maps 53 -> 52 for libraries that require a consistent
measure for weeks.
}
\examples{
  x <- seq( ymd(20011231), ymd(20301231), "year" )
  lubridate::week(x)
  week52(x)

}
