% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesterday.R
\name{yesterday}
\alias{last_hour}
\alias{last_month}
\alias{next_hour}
\alias{next_month}
\alias{tomorrow}
\alias{yesterday}
\title{Reference Dates and Times from Now}
\usage{
yesterday(tzone = NULL)

tomorrow(tzone = NULL)

last_month(tzone = NULL)

next_month(tzone = NULL)

last_hour(tzone = NULL)

next_hour(tzone = NULL)
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like
the current time in. tzone defaults to your computer's system timezone. You
can retrieve the current time in the Universal Coordinated Time (UTC) with
now("UTC").}
}
\description{
yesterday, tomorrow, last month and next month based on the current datetime
}
\details{
These functions provide commonly used dates relative to the current date and
time (\code{now}). They give the floor for the specified time period.
}
\examples{

  yesterday()
  tomorrow()

  last_month()
  next_month()

  last_hour()
  next_hour()
 #  ( now() - ddays(1) ) > yesterday()

}
\seealso{
\code{\link[lubridate]{now}} \cr
  \code{\link[lubridate]{floor_date}} \cr
}

