% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-rounding.R
\name{dayify}
\alias{ceiling_day}
\alias{ceiling_hour}
\alias{ceiling_minute}
\alias{ceiling_month}
\alias{ceiling_second}
\alias{ceiling_week}
\alias{ceiling_year}
\alias{dayify}
\alias{floor_day}
\alias{floor_hour}
\alias{floor_minute}
\alias{floor_month}
\alias{floor_second}
\alias{floor_week}
\alias{floor_year}
\alias{monthify}
\alias{yearify}
\title{Date rounding}
\usage{
dayify(x)

monthify(x)

yearify(x)

floor_year(...)

floor_month(...)

floor_week(...)

floor_day(...)

floor_hour(...)

floor_minute(...)

floor_second(...)

ceiling_year(...)

ceiling_month(...)

ceiling_week(...)

ceiling_day(...)

ceiling_hour(...)

ceiling_minute(...)

ceiling_second(...)
}
\arguments{
\item{x}{vector of date time objects}

\item{...}{arguments passed to other functions}
}
\description{
Function to round dates to various units
}
\details{
The 'date'-ify functions round each date down to the beginning of that
period using the \code{\link{floor_date}} function. This
allows the continued use of date functions.
}
\note{
TODO
  - convert automatically from character
}
\seealso{
\code{\link{floor_date}}
}

