% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-differences.R
\name{diffdays}
\alias{diffdays}
\alias{diffweeks}
\alias{diffyears}
\title{Difference between dates expressed in days or weeks}
\usage{
diffdays(x, y)

diffweeks(x, y)

diffyears(x, y)
}
\arguments{
\item{x}{the first date/time object}

\item{y}{the second date/time object}
}
\description{
Difference between dates expressed in days or weeks
}
\references{
\url{http://stackoverflow.com/questions/15569333/r-get-date-difference-in-years-floating-point}
}
\seealso{
\code{\link[base]{difftime}}
}

