% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-memoised.R
\name{ceiling_date_memo}
\alias{ceiling_date_memo}
\alias{floor_date_memo}
\alias{ymd_memo}
\title{Memoised functions}
\usage{
ceiling_date_memo(x, unit = "seconds", change_on_boundary = NULL,
  week_start = getOption("lubridate.week.start", 7))

floor_date_memo(x, unit = "seconds",
  week_start = getOption("lubridate.week.start", 7))

ymd_memo(..., quiet = FALSE, tz = NULL, locale = Sys.getlocale("LC_TIME"),
  truncated = 0)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Memoized versions of functions that can benefit from memoisation.
}
\details{
These functions have been memoised for speed and cache the return
values associated with their arguments.
}
