% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/date-memoised.R
\name{ceiling_date_memo}
\alias{ceiling_date_memo}
\alias{floor_date_memo}
\alias{ymd_memo}
\title{Memoised functions}
\usage{
ceiling_date_memo(...)

floor_date_memo(...)

ymd_memo(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Memoized versions of functions that can benefit from memoisation.
}
\details{
These functions have been memoised for speed and cache the return
values associated with their arguments.
}

