% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch-origin.R, R/origin.R
\docType{data}
\name{origin}
\alias{epoch_origin}
\alias{origin}
\alias{origin_Date}
\alias{origin_POSIXct}
\alias{origin_POSIXlt}
\alias{origin_chron}
\alias{origin_epoch}
\title{Get origin for a time series class}
\format{An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.}
\usage{
origin(x, ...)

epoch_origin

origin_epoch

origin_POSIXct

origin_POSIXlt

origin_chron

origin_Date

origin(x, ...)
}
\arguments{
\item{x}{time series object}

\item{...}{addition arguments}

\item{tzone}{string; time zone abbreviation}

\item{x}{time series object}

\item{...}{addition arguments}

\item{tzone}{string; time zone abbreviation}
}
\value{
A one element vector with the class of \code{x}, denoting the

A one element vector with the class of \code{x}, denoting the
}
\description{
Get origin for a time series class

Origin is the date-time for 1970-01-01 UTC in POSIXct format. This date-time
is the origin for the numbering system used by POSIXct, POSIXlt, chron, and
Date classes.

Return the origin for a time-series class
}
\details{
This is a copy of \code{\link[lubridate]{origin}}.
}
\examples{
epoch_origin

}
\seealso{
lubridate::origin

\code{\link{origins}}
}
\keyword{chron}
\keyword{data}
\keyword{datasets}

