% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesterday.R
\name{yesterday}
\alias{tomorrow}
\alias{yesterday}
\title{yesterday and tomorrow}
\usage{
yesterday(tzone = NULL)

tomorrow(tzone = "")
}
\arguments{
\item{tzone}{a character vector specifying which time zone you would like
the current time in. tzone defaults to your computer's system timezone. You
can retrieve the current time in the Universal Coordinated Time (UTC) with
now("UTC").}
}
\description{
This same time yesterday
}
\examples{

  yesterday()
  tomorrow()

 #  ( now() - ddays(1) ) > yesterday()

}
\seealso{
\code{\link[lubridate]{here}}
}

