% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym}
\alias{ym}
\title{Representative Dates as the first day in the period}
\usage{
ym(x, ...)
}
\arguments{
\item{x}{vector; values that represent months in ym form}

\item{...}{arguments passed to ymd}
}
\value{
POSIXct
}
\description{
Represent a data as the first day of the month of an interval.
}
\details{
Parses a ym date string by setting the date/time to the first moment of the
month. If \code{x} is already a POSIX date, \code{floor_month(x)} is performed.
}
\examples{
  months <- c('200101','200102', '200103' )
  ym(months)

}
\seealso{
\code{\link[lubridate]{yday}}
}
